Move GeneratedImage to Images::GenerateSocialmage
