+++ spec/helpers/social_image_helper_spec.rb
@@ -24,7 +24,7 @@ describe SocialImageHelper do
 
       url = helper.user_social_image_url(user)
 
-      expect(url).to eq GeneratedImage.new(user).social_image
+      expect(url).to eq Images::GenerateSocialImage.call(user)
     end
 
     it "returns social preview path for newer decorated users" do
@@ -56,7 +56,7 @@ describe SocialImageHelper do
 
       url = helper.article_social_image_url(article)
 
-      expect(url).to eq GeneratedImage.new(article).social_image
+      expect(url).to eq Images::GenerateSocialImage.call(article)
     end
 
     it "returns social preview path for newer decorated articles" do
